# lab2/payment-auth-system/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateCardRequest:
    properties:
      balance:
        type: integer
      expire_date:
        type: string
      is_blocked:
        type: boolean
      key_id:
        type: integer
      number:
        type: string
      owner_name:
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      token:
        type: string
    type: object
  handlers.UpdateCardRequest:
    properties:
      balance:
        type: integer
      expire_date:
        type: string
      is_blocked:
        type: boolean
      key_id:
        type: integer
      number:
        type: string
      owner_name:
        type: string
    type: object
...
  models.Card:
    properties:
      balance:
        format: int64
        type: integer
      createdAt:
        type: string
      expireDate:
        type: string
      id:
        format: int64
        type: integer
      isBlocked:
        type: boolean
      keyID:
        format: int64
        type: integer
      number:
        type: string
      ownerName:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API сервера авторизации платежей транспортными картами
  title: Payment Auth System API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация по логину и паролю, возврат JWT
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Вход пользователя
      tags:
      - auth
  /cards:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Card'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить список карт
      tags:
      - cards
    post:
      consumes:
      - application/json
      parameters:
      - description: Create card request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать карту
      tags:
      - cards
  /cards/{id}:
    delete:
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить карту
      tags:
      - cards
    get:
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Card'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить карту по ID
      tags:
      - cards
    put:
      consumes:
      - application/json
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update card request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить карту
      tags:
      - cards
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"